test(recipe): relocate DirectionStepsTest to public module
The Content() refactor moved RecipeDetailScreen (and the internal
directionSteps helper) from recipe/core/impl into recipe/core/public,
but left DirectionStepsTest behind in impl/commonTest where the now
cross-module internal helper is no longer visible, breaking test
compilation on CI.

Move the test into recipe/core/public/commonTest under the matching
package so it can see the internal helper, and enable testing on the
public module.
