+++ client/recipe/core/public/build.gradle.kts
@@ -29,4 +29,7 @@ kotlin {
 
 compose { resources { publicResClass = true } }
 
-plusLibrary { namespace = "com.plusmobileapps.chefmate.recipe.core" }
+plusLibrary {
+    namespace = "com.plusmobileapps.chefmate.recipe.core"
+    enableTesting = true
+}
