+++ client/recipe/core/public/src/commonTest/kotlin/com/plusmobileapps/chefmate/recipe/core/detail/DirectionStepsTest.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.chefmate.recipe.core.impl.detail.ui
+package com.plusmobileapps.chefmate.recipe.core.detail
 
 import io.kotest.matchers.shouldBe
 import kotlin.test.Test
